Owner repo and output file no longer needed

+++ optcd.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
-if [ "$#" -le 3 ]; then
-  echo "Usage: run.sh <input_yaml_filename> <output_yaml_filename> <owner> <repo> [output_file]"
+if [ "$#" -le 1 ]; then
+  echo "Usage: run.sh <input_yaml_filename> <output_yaml_filename>"
   exit 1
 fi
 
@@ -12,9 +12,43 @@ pip install -r requirements.txt > /dev/null 2>&1
 currentDir=$(pwd)
 input_yaml_filename=$1
 output_yaml_filename=$2
-owner=$3
-repo=$4
-initial_output_file=$5
+
+# go to the input_yaml_filename directory, and find out owner and repo for pushing the modified YAML file
+project_dir=$(dirname "$input_yaml_filename")
+cd "$project_dir"
+if [ $? -ne 0 ]; then
+  echo "Error: Unable to change directory to $project_dir"
+  exit 1
+fi
+url=$(git config --get remote.origin.url)
+
+if [[ $url =~ github.com[:/](.*)/(.*)(\.git)?$ ]]; then
+  owner="${BASH_REMATCH[1]}"
+  repo="${BASH_REMATCH[2]}" # repo name contains the .git extension
+  # remove the .git extension if exists
+  if [[ $repo == *.git ]]; then
+    repo="${repo%.git}"
+  fi
+else
+  echo "Could not parse owner and repo from URL: $url"
+  exit 1
+fi
+
+cd "$currentDir"
+if [ $? -ne 0 ]; then
+  echo "Error: Unable to change directory back to $currentDir"
+  exit 1
+fi
+
+# if $3 and $4 are not empty, then use them as owner and repo
+# if [ -n "$3" ] && [ -n "$4" ]; then
+#   owner=$3
+#   repo=$4
+# fi
+
+# owner=$3
+# repo=$4
+initial_output_file="out.txt"
 fixed_output_file="temp_out.txt"
 
 rm $initial_output_file > /dev/null 2>&1
